+++ taiga/export_import/api.py
@@ -104,7 +104,7 @@ class ProjectImporterViewSet(mixins.ImportThrottlingPolicyMixin, CreateModelMixi
         is_private = data.get('is_private', False)
         total_memberships = len([m for m in data.get("memberships", []) if m.get("email", None) != data["owner"]])
         total_memberships = total_memberships + 1  # 1 is the owner
-        (enough_slots, error_message) = users_services.has_available_slot_for_import_new_project(
+        (enough_slots, error_message) = users_services.has_available_slot_for_new_project(
             self.request.user,
             is_private,
             total_memberships
@@ -344,7 +344,7 @@ class ProjectImporterViewSet(mixins.ImportThrottlingPolicyMixin, CreateModelMixi
         total_memberships = len([m for m in dump.get("memberships", [])
                                             if m.get("email", None) != dump["owner"]])
         total_memberships = total_memberships + 1 # 1 is the owner
-        (enough_slots, error_message) = users_services.has_available_slot_for_import_new_project(
+        (enough_slots, error_message) = users_services.has_available_slot_for_new_project(
             user,
             is_private,
             total_memberships