+++ taiga/projects/apps.py
@@ -58,6 +58,7 @@ def connect_memberships_signals():
                               sender=apps.get_model("projects", "Membership"),
                               dispatch_uid='create-notify-policy')
 
+
 def disconnect_memberships_signals():
     signals.pre_delete.disconnect(sender=apps.get_model("projects", "Membership"),
                                   dispatch_uid='membership_pre_delete')
@@ -79,7 +80,6 @@ def disconnect_us_status_signals():
                                  dispatch_uid="try_to_close_or_open_user_stories_when_edit_us_status")
 
 
-
 ## Tasks Statuses Signals
 
 def connect_task_status_signals():
@@ -92,7 +92,6 @@ def connect_task_status_signals():
 def disconnect_task_status_signals():
     signals.post_save.disconnect(sender=apps.get_model("projects", "TaskStatus"),
                                  dispatch_uid="try_to_close_or_open_user_stories_when_edit_task_status")
-
 
 
 class ProjectsAppConfig(AppConfig):